backend/requirements.txt
fastapi
uvicorn
python-chess
websockets
